+++ slidershow/operation.js
@@ -24,6 +24,7 @@ class Operation {
      */
     _button(group_name) {
         const $group = $("<div/>", { "data-hotkey-group": group_name }).appendTo(this.playback.hud.$hud_menu)
+        /**@param {Button} button */
         return ([hotkey, symbol, hint, fn, role = null]) => [hotkey,
             $("<button/>", { "title": hint, "data-hotkey": hotkey, "html": symbol, "data-role": role })
                 .on("click", fn)
@@ -31,49 +32,94 @@ class Operation {
         ]
     }
 
+    /**
+     *
+     * @typedef {[hotkey: string|string[], symbol: string, hint: string, fn: Function, role?: string|null]} Button
+     * @typedef {[hotkey: string|string[], hint: string, fn: Function]} Shortcut
+     * @typedef {[hint: string, fn: Function, isEnabled: Function, group_name: string]} Command
+     *
+     * Register a shortcut in the command palette, as hotkeys, and possibly as a GUI button
+     *
+     * @param {string} group_name
+     * @param {?Shortcut[]} shortcuts Shortcuts without a GUI button
+     * @param {Button[]} buttons Shortcuts with a GUI button.
+     * @param {?Command[]} commands Commands without shortcuts.
+     * @returns
+     */
+    _group(group_name, shortcuts, buttons, commands = null) {
+        if (!shortcuts) {
+            shortcuts = []
+        } else {
+            // separate shortcuts with multiple hotkeys into two shortcuts
+            shortcuts = shortcuts.flatMap(shortcut => {
+                const [hotkey, hint, fn] = shortcut
+                if (Array.isArray(hotkey)) {
+                    return hotkey.map(hk => [hk, hint, fn])
+                }
+                return [shortcut]
+            })
+        }
+
+        // separate buttons with multiple hotkeys:
+        // first hotkey will become button, other shortcuts
+        // Ex. ['/', '?'] -> Button with '/', alternate shortcut with '/'
+        const normalizedButtons = buttons.map(button => {
+            const [hotkey, symbol, hint, fn, role = null] = button
+            if (Array.isArray(hotkey)) {
+                const [primaryHotkey, ...otherHotkeys] = hotkey
+                otherHotkeys.forEach(hk => shortcuts.push([hk, hint, fn]))
+                return [primaryHotkey, symbol, hint, fn, role]
+            }
+            return button
+        })
+
+        const group = wh.group(group_name, normalizedButtons.map(this._button(group_name)))
+        if (shortcuts.length) {
+            wh.group(group_name, shortcuts)
+        }
+        for (const hotkey of group) {
+            const callback = hotkey.element ? () => hotkey.element.click() : hotkey.action
+            this.playback.hud.palette.register(hotkey.hint, callback, () => hotkey.enabled, hotkey.getClue(), group_name)
+        }
+        for (const cmd of commands || []) {
+            this.playback.hud.palette.register(cmd[0], cmd[1], cmd[2], null, group_name + " > " + cmd[3])
+        }
+        return group
+    }
+
     taggingInit() {
         const pl = this.playback
-        return wh.group("Tagging", [ // hotkey duplicates without GUI buttons
-            ["Digit0", "Tag 0", () => pl.frame.set_tag(null)],
-            ["Digit1", "Tag 1", () => pl.frame.set_tag(1)],
-            ["Digit2", "Tag 2", () => pl.frame.set_tag(2)],
-            ["Digit3", "Tag 3", () => pl.frame.set_tag(3)],
-            ["Digit4", "Tag 4", () => pl.frame.set_tag(4)],
-            ["Digit5", "Tag 5", () => pl.frame.set_tag(5)],
-            ["Digit6", "Tag 6", () => pl.frame.set_tag(6)],
-            ["Digit7", "Tag 7", () => pl.frame.set_tag(7)],
-            ["Digit8", "Tag 8", () => pl.frame.set_tag(8)],
-            ["Digit9", "Tag 9", () => pl.frame.set_tag(9)],
-            ...[
+        return this._group("Tagging", [
+            [["Numpad4", "Digit4"], "Tag 4", () => pl.frame.set_tag(4)],
+            [["Numpad5", "Digit5"], "Tag 5", () => pl.frame.set_tag(5)],
+            [["Numpad6", "Digit6"], "Tag 6", () => pl.frame.set_tag(6)],
+            [["Numpad7", "Digit7"], "Tag 7", () => pl.frame.set_tag(7)],
+            [["Numpad8", "Digit8"], "Tag 8", () => pl.frame.set_tag(8)],
+            [["Numpad9", "Digit9"], "Tag 9", () => pl.frame.set_tag(9)],
+            ["Alt+Numpad0", "Tag 10", () => pl.frame.set_tag(10)],
+            ["Alt+Numpad1", "Tag 11", () => pl.frame.set_tag(11)],
+            ["Alt+Numpad2", "Tag 12", () => pl.frame.set_tag(12)],
+            ["Alt+Numpad3", "Tag 13", () => pl.frame.set_tag(13)],
+            ["Alt+Numpad4", "Tag 14", () => pl.frame.set_tag(14)],
+            ["Alt+Numpad5", "Tag 15", () => pl.frame.set_tag(15)],
+            ["Alt+Numpad6", "Tag 16", () => pl.frame.set_tag(16)],
+            ["Alt+Numpad7", "Tag 17", () => pl.frame.set_tag(17)],
+            ["Alt+Numpad8", "Tag 18", () => pl.frame.set_tag(18)],
+            ["Alt+Numpad9", "Tag 19", () => pl.frame.set_tag(19)],
+        ],
+            [
                 ["Alt+Shift+g", "🔀", "Group frames according to their tag", () => pl.section_controller.group()],
-                ["Numpad0", "⛔", "Tag 0", () => pl.frame.set_tag(null)],
-                ["Numpad1", "1", "Tag 1", () => pl.frame.set_tag(1)],
-                ["Numpad2", "2", "Tag 2", () => pl.frame.set_tag(2)],
-                ["Numpad3", "3", "Tag 3", () => pl.frame.set_tag(3)],
-                ["Numpad4", "4", "Tag 4", () => pl.frame.set_tag(4)],
-                ["Numpad5", "5", "Tag 5", () => pl.frame.set_tag(5)],
-                ["Numpad6", "6", "Tag 6", () => pl.frame.set_tag(6)],
-                ["Numpad7", "7", "Tag 7", () => pl.frame.set_tag(7)],
-                ["Numpad8", "8", "Tag 8", () => pl.frame.set_tag(8)],
-                ["Numpad9", "9", "Tag 9", () => pl.frame.set_tag(9)],
-                ["Alt+Numpad0", "10", "Tag 10", () => pl.frame.set_tag(10)],
-                ["Alt+Numpad1", "11", "Tag 11", () => pl.frame.set_tag(11)],
-                ["Alt+Numpad2", "12", "Tag 12", () => pl.frame.set_tag(12)],
-                ["Alt+Numpad3", "13", "Tag 13", () => pl.frame.set_tag(13)],
-                ["Alt+Numpad4", "14", "Tag 14", () => pl.frame.set_tag(14)],
-                ["Alt+Numpad5", "15", "Tag 15", () => pl.frame.set_tag(15)],
-                ["Alt+Numpad6", "16", "Tag 16", () => pl.frame.set_tag(16)],
-                ["Alt+Numpad7", "17", "Tag 17", () => pl.frame.set_tag(17)],
-                ["Alt+Numpad8", "18", "Tag 18", () => pl.frame.set_tag(18)],
-                ["Alt+Numpad9", "19", "Tag 19", () => pl.frame.set_tag(19)]
-            ].map(this._button("Tagging"))
-        ]).toggle(pl.tagging_mode)
+                [["Numpad0", "Digit0"], "⛔", "Tag 0", () => pl.frame.set_tag(null)],
+                [["Numpad1", "Digit1"], "1", "Tag 1", () => pl.frame.set_tag(1)],
+                [["Numpad2", "Digit2"], "2", "Tag 2", () => pl.frame.set_tag(2)],
+                [["Numpad3", "Digit3"], "3", "Tag 3", () => pl.frame.set_tag(3)],
+            ]).toggle(pl.tagging_mode)
     }
 
     editingInit() {
         const pl = this.playback
         const cc = pl.changes
-        return wh.group("Editing", [
+        return this._group("Editing", null, [
             ["Alt+d", "🐐🐐", "Duplicate frame", () => {
                 const f = pl.frame
                 f.leave() // assure the frame is fully unloaded so that we do not inherit ex: common wzoom object
@@ -139,16 +185,16 @@ class Operation {
                 }
             }],
             ["Escape", "✔️", "Stop editing", () => $(":focus").blur(), "stop-editing"]
-        ].map(this._button("Editing"))).toggle(pl.editing_mode)
+        ]).toggle(pl.editing_mode)
     }
 
     playthroughInit() {
         const pl = this.playback
-        return wh.group("Playthrough", [
+        return this._group("Playthrough", [
             ["Ctrl+Space", "Play/Pause", () => { // undocumented
                 pl.play_pause(!pl.moving)
-            }],
-            ...[
+            }]],
+            [
                 ["Home", "⏮", "Go to the first", () => pl.goToFrame(0)],
                 ["Alt+PageUp", "◀◀", "Prev section", () => pl.previousSection()],
                 ["Shift+PageUp", "◀", "Prev frame", () => pl.previousFrame()],
@@ -197,13 +243,13 @@ class Operation {
                 ["Shift+PageDown", "▶", "Next frame", () => pl.nextFrame()],
                 ["Alt+PageDown", "▶▶", "Next section", () => pl.nextSection()],
                 ["End", "⏭", "Go to end", () => pl.goToFrame(pl.$articles.length - 1)],
-            ].map(this._button("Playthrough"))]).disable()
+            ]).disable()
     }
 
     generalInit() {
         const pl = this.playback
-        return wh.group("General", [
-            ...[
+        return this._group("General", null,
+            [
                 ["m", "🗺", "Toggle hud map", () => pl.hud_map.toggle(true), "not-video"],
                 ["i", "ℹ", "Toggle file info", () => $("#hud-fileinfo").toggle()],
                 ["z", "🔍", "Photo or video zoom (cycle)", () => zoom()],
@@ -221,7 +267,7 @@ class Operation {
                         }]
                     })
                 }]
-            ].map(this._button("General"))]).disable()
+            ]).disable()
 
         function zoom(little) {
             const wzoom = pl.frame.$actor?.data("wzoom")
@@ -243,8 +289,8 @@ class Operation {
 
     mediaInit() {
         const pl = this.playback
-        return wh.group("Media", [
-            ...[
+        return this._group("Media", null,
+            [
                 ["NumpadAdd", "🐰", "Faster video", () => playback_change(0.1), "only-video"],
                 ["NumpadSubtract", "🐢", "Slower video", () => playback_change(-0.1), "only-video"],
                 ["Alt+m", "🔇", "Toggle muted", () => {
@@ -253,7 +299,7 @@ class Operation {
                 ["Shift+r", "⤿", "Rotate left", () => rotate(-5)],
                 ["Shift+Alt+r", "⤾", "Rotate right", () => rotate(5)],
                 ["r", "⊾", "Rotate right 90°", () => rotate(90)],
-            ].map(this._button("Media"))]).disable()
+            ]).disable()
 
         function act() {
             return pl.frame.$actor
@@ -275,56 +321,58 @@ class Operation {
 
     propertiesInit() {
         const pl = this.playback
-        return wh.group("Properties", [
-            ...[
+        return this._group("Properties", null,
+            [
                 // NOTE: If you change the frame with the property panel closed,
                 // the shortcut still triggers the button on the old frame
                 // causing the browser to go back and register unzoomed position as a data-point.
                 ["Alt+s", "📸", "Add step point", () =>
                     $(".hud-point", pl.hud.$hud).eq(1).trigger("click")
                 ],
-            ].map(this._button("Properties"))]).disable()
-
-        function act() {
-            return pl.frame.$actor
-        }
+            ]).disable()
     }
 
     gridInit() {
         const pl = this.playback
-        return wh.group("Grid", [
-            ...[
-                ["Enter", "✓", "Enter the frame (hides the grid)", () => this.playback.hud.toggle_grid()],
+        return this._group("Grid", [
+            ["Enter", "Enter the frame (hides the grid)", () => this.playback.hud.toggle_grid()],
 
-                // NOTE we may implement selections. In that case, these shortcuts should be hidden or moved to a command palette. Too much of them!
-                // ["Shift+ArrowRight", "select right", "Add right frame to selection", () => pl.section_controller.moveFrame(pl.index, pl.index+1, false)],
-                ["Ctrl+ArrowUp", "△⇵", "Move up", () => pl.section_controller.moveFrame(pl.index, pl.hud.grid.getFrameIndexInNextRow(-1), true)],
-                ["Ctrl+ArrowDown", "▽⇵", "Move down", () => pl.section_controller.moveFrame(pl.index, pl.hud.grid.getFrameIndexInNextRow(1), false)],
-                ["Ctrl+ArrowLeft", "◁⇵", "Move left", () => pl.section_controller.moveFrame(pl.index, pl.index - 1, true)],
-                ["Ctrl+ArrowRight", "▷⇵", "Move right", () => pl.section_controller.moveFrame(pl.index, pl.index + 1, false)],
+            // NOTE we may implement selections. In that case, these shortcuts should be hidden or moved to a command palette. Too much of them!
+            // ["Shift+ArrowRight", "select right", "Add right frame to selection", () => pl.section_controller.moveFrame(pl.index, pl.index+1, false)],
+            ["Ctrl+ArrowUp", "Move up", () => pl.section_controller.moveFrame(pl.index, pl.hud.grid.getFrameIndexInNextRow(-1), true)],
+            ["Ctrl+ArrowDown", "Move down", () => pl.section_controller.moveFrame(pl.index, pl.hud.grid.getFrameIndexInNextRow(1), false)],
+            ["Ctrl+ArrowLeft", "Move left", () => pl.section_controller.moveFrame(pl.index, pl.index - 1, true)],
+            ["Ctrl+ArrowRight", "Move right", () => pl.section_controller.moveFrame(pl.index, pl.index + 1, false)],
 
-                ["ArrowUp", "△", "Go up", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextRow(-1))],
-                ["ArrowDown", "▽", "Go down", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextRow(1))],
-                ["ArrowLeft", "◁", "Go left", () => pl.previousFrame()], // normally, left arrow triggers next step but this would block the grid, we need next frame
-                ["ArrowRight", "▷", "Go right", () => pl.nextFrame()],
+            ["ArrowUp", "Go up", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextRow(-1))],
+            ["ArrowDown", "Go down", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextRow(1))],
+            ["ArrowLeft", "Go left", () => pl.previousFrame()], // normally, left arrow triggers next step but this would block the grid, we need next frame
+            ["ArrowRight", "Go right", () => pl.nextFrame()],
 
-                ["PageUp", "⮅", "Page up", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextPage(-1))],
-                ["PageDown", "⮇", "Page down", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextPage(1))],
+            ["PageUp", "Page up", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextPage(-1))],
+            ["PageDown", "Page down", () => pl.goToFrame(pl.hud.grid.getFrameIndexInNextPage(1))],
+        ],
+            [
+                ["Alt+?", "✥", "Navigation help", () =>
+                    new $.Zebra_Dialog("Use arrows to navigate.<br>Ctrl+Arrow to move.<br>PageUp/Down.<br>Home/End.<br>Enter to access the frame.", {
+                        title: "Grid navigation help",
+                        buttons: false
+                    })],
                 ["NumpadAdd", "+", "More thumbnails on a row", () => pl.hud.grid.changeColumnsCount(1)],
                 ["NumpadSubtract", "-", "Less thumbnails on a row", () => pl.hud.grid.changeColumnsCount(-1)],
 
-            ].map(this._button("Grid"))]).disable()
+            ], pl.hud.grid.getHotkeys()).disable()
     }
 
     switchesInit() {
         const pl = this.playback
-        return wh.group("Switches", [
+        return this._group("Switches", [
             ["Ctrl+Alt+d", "Debug", () => { // undocumented
                 const zoom = $main.css("zoom")
                 $main.css({ "zoom": zoom == "1" ? "0.05" : "1" })
                 pl.debug = !pl.debug
-            }],
-            ...[
+            }]],
+            [
                 ["Ctrl+Alt+z", "⟲", "Undo change", () => pl.changes.undo(), "undo"],
                 ["Ctrl+Alt+Shift+z", "⟳", "Redo change", () => pl.changes.redo(), "redo"],
                 ["j", "&#127895;", "Thumbnails", () => pl.hud.toggle_thumbnails()],
@@ -356,18 +404,18 @@ class Operation {
                     pl.hud.info(`Tagging mode ${pl.tagging_mode ? "enabled" : "disabled."}`)
                     pl.session.store()
                 }]
-            ].map(this._button("Switches"))
-        ]).disable()
+            ]).disable()
     }
 
     globalInit() {
         const menu = this.playback.menu
-        return wh.group("Global", [
-            ["Escape", "☰", "Show menu", () => this.playback.hud.$playback_icon.click()],
+        return this._group("Global", null, [
+            ["Escape", "☰", "Toggle menu", () => this.playback.hud.toggleMenu()],
+            [["/", "?"], "🎨", "Command palette", () => this.playback.hud.palette.focus()],
             ["Alt+m", "🧰", "Show splashscreen", () => menu.stop_playback()],
             ["Alt+w", "&#127916;", "Auxiliary window", () => menu.aux_window.open()],
             ['Ctrl+s', "&#128190;", "Export", () => menu.export.export_dialog()],
             ['F1', "&#9432;", "Help", () => menu.help()],
-        ].map(this._button("Global")))
+        ])
     }
 }